requirements.txt
langgraph
langchain
langsmith
python-dotenv
langchain-groq
langchain_tavily
langgraph-cli[inmem]
pymupdf
langchain-core
langchain-community
langchain-openai
transformers
torch
pillow
numpy
scikit-learn
faiss-cpu
langchain-google-genai
google-generativeai